➕ Component for item form
